Collapsible SingleMetadataResult, w/ updated test
